requirements.txt
torch
numpy
tensorboard
Pillow
pandas
